fix(control): circular dependencies in the isControl functions
